imported updated locomotion package

+++ FallingCubePrototype/Assets/Resources/Invector-3rdPersonController_LITE/Scripts/ObjectDamage/vDamage.cs
@@ -5,7 +5,7 @@ namespace Invector
     public class vDamage
     {
         [Tooltip("Apply damage to the Character Health")]
-        public int damageValue = 15;
+        public float damageValue = 15;
         [Tooltip("How much stamina the target will lost when blocking this attack")]
         public float staminaBlockCost = 5;
         [Tooltip("How much time the stamina of the target will wait to recovery")]
